Pingu Bot Rewrite with better i18n support (#44)
* Dependa: Añadir paquete string-placeholder para mejor las traducciones.

* Comandos: Convertir todos los mensajes de respuesta a mensajes enriquecidos. Añadir comando help básico.

+++ tools/security/clear.js
@@ -1,49 +1,56 @@
-const { Permissions } = require('discord.js')
+const { Permissions, MessageEmbed } = require('discord.js')
+const genericMessages = require('../../modules/genericMessages')
+const getLocales = require('../../modules/getLocales')
 
 module.exports = {
   name: 'clear',
-  execute (args, client, con, contenido, message, result) {
-    const i18n = require(`../../i18n/${result[0].guild_language}.json`).tools.security.clear
-    if (message.member.permissions.has([Permissions.FLAGS.MANAGE_MESSAGES, Permissions.FLAGS.KICK_MEMBERS, Permissions.FLAGS.BAN_MEMBERS]) || message.member.permissions.has([Permissions.FLAGS.ADMINISTRATOR])) {
-      if (result[0].moderator_enabled !== 0) {
-        if (isNaN(args[0])) {
+  execute (args, client, con, locale, message, result) {
+    if (result[0].moderator_enabled !== 0) {
+      if (message.member.permissions.has([Permissions.FLAGS.MANAGE_MESSAGES, Permissions.FLAGS.KICK_MEMBERS, Permissions.FLAGS.BAN_MEMBERS]) || message.member.permissions.has([Permissions.FLAGS.ADMINISTRATOR])) {
+        args[0] = parseInt(args[0].replace(/\D/g, ''))
+        if (args[0] && parseInt(args[0]) <= 100) {
           if (message.mentions.users.first()) {
-            message.channel.messages.fetch({
-              // Eliminar del message.content el comando y las menciones del comando.
-              limit: args[0]
-            }).then((messages) => {
+            message.delete()
+            message.channel.messages.fetch().then((messages) => {
               const botMessages = []
-              message.channel.bulkDelete(1, true)
-              messages.filter(m => m.author.id === message.mentions.users.first().id).forEach(msg => botMessages.push(msg))
-              message.channel.bulkDelete(botMessages).then((_message) => {
-                message.channel.send(`:broom: \`${_message.size - 1}\` ${i18n.success}`).then(msg => msg.delete({
-                  timeout: 2500
-                }))
+              let i = 0
+              messages.filter(m => m.author.id === message.mentions.users.first().id).forEach(msg => {
+                if (i < args[0]) {
+                  i++
+                  botMessages.push(msg)
+                }
+              })
+              message.channel.bulkDelete(botMessages, true).then((_message) => {
+                const sent = new MessageEmbed().setColor('#28A745').setDescription(`:broom: ${getLocales(locale, 'CLEAR_EMBED_TITLE', { BROOM: _message.size })}`)
+                message.channel
+                  .send({ embeds: [sent] }).then((sent) => {
+                    setTimeout(() => {
+                      sent.delete()
+                    }, 2500)
+                  })
               })
             })
           } else {
-            if (args[0]) {
-              const i = parseInt(args[0])
-
-              message.channel.bulkDelete(i + 1, true)
-                .then((_message) => {
-                  message.channel
-                    .send(`:broom: \`${_message.size - 1}\` ${i18n.success}`).then((sent) => {
-                      setTimeout(() => {
-                        sent.delete()
-                      }, 2500)
-                    })
-                })
-            } else {
-              message.channel.send(`<:win_information:876119543968305233> ${i18n.missing_arg} \`${result[0].guild_prefix}del <cantidad> \``)
-            }
+            message.delete()
+            message.channel.bulkDelete(args[0], true)
+              .then((_message) => {
+                const sent = new MessageEmbed().setColor('#28A745').setDescription(`:broom: ${getLocales(locale, 'CLEAR_EMBED_TITLE', { BROOM: _message.size })}`)
+                message.channel
+                  .send({ embeds: [sent] }).then((sent) => {
+                    setTimeout(() => {
+                      sent.delete()
+                    }, 2500)
+                  })
+              })
           }
         } else {
-          message.reply('PLACEHOLDER_ARGS0_NOT_INT')
+          genericMessages.Info.help(message, locale, `${result[0].guild_prefix}clear <number <= 100> (@user)`)
         }
+      } else {
+        genericMessages.Error.permerror(message, locale)
       }
     } else {
-      message.channel.send(`<:pingu_false:876103413526564924> ${i18n.permerror}`)
+      genericMessages.Error.no_avaliable(message, locale)
     }
   }
 }
